refactor(hardpoints): fold new.vue into the Hardpoints index component
With old.vue gone, new.vue was the only implementation and index.vue was a
pass-through wrapper. Merge them so the component is just Hardpoints, and
rename new.scss to match.
