+++ app/frontend/frontend/components/Models/Hardpoints/index.scss
